Move dataset api code to api package for consistency
